\name{rorranking-package}
\alias{rorranking-package}
\alias{rorranking}
\docType{package}
\title{
Implementations of ror ranking methods
}
\description{
This package contains an implementations of ROR ranking methods.  
}
\details{
\tabular{ll}{
Package: \tab rorranking\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-05\cr
License: \tab GPL-3\cr
}
}
\author{
Maintainer: Pawel Rychly <pawelrychly@gmail.com>
}

